Add files via upload
I have used this in production but it has problems.  I believe the 500Hz update rate of the controls is too much for the Arduino to handle and signals are changing while interrupt handling code is still running.  This causes erroneous readings and the tractor twitches occasionally under local control.